mytrains/requirements.txt
attr==0.3.2
attrs==21.2.0
bpython==0.23
ConfigParser==5.3.0
cryptography==3.4.8
cx_Oracle==8.3.0
Cython==0.29.24
djangorestframework==3.13.1
dl==0.1.0
docutils==0.17.1
HTMLParser==0.0.2
ipdb==0.13.9
ipython==8.5.0
ipywidgets==7.6.5
jnius==1.1.0
keyring==23.1.0
lockfile==0.12.2
MySQL-python==1.2.5
numpy==1.20.3
Pillow==9.2.0
protobuf==4.21.6
psycopg2==2.9.3
pylibmc==1.6.3
pymemcache==3.5.2
pyOpenSSL==22.0.0
pytz==2021.3
pywatchman==1.4.1
railroad==0.5.0
Sphinx==4.2.0
tblib==1.7.0
toml==0.10.2
tornado==6.1
trove_classifiers==2022.8.31
truststore==0.5.0
xmlrpclib==1.0.1
